arm: aspeed: Add AST2600 Tacoma device tree
This describes the ethernet, spi nor and emmc devices on the ast2600
based Tacoma board.

The ethernet configuration is for NCSI on the third MAC with internal
RCLK.

WDT1 is used for fan control on this platform so it is left disabled.
WDT2 is enabled for reboot.

The FMC (flash memory controller) is configured to use Quad IO for the
BMC flash.

Signed-off-by: Joel Stanley <joel@jms.id.au>